+++ src/Aws/DynamoDb/Exception/LimitExceededException.php
@@ -17,6 +17,6 @@
 namespace Aws\DynamoDb\Exception;
 
 /**
- * The number of concurrent table requests (cumulative number of tables in the CREATING, DELETING or UPDATING state) exceeds the maximum allowed of 20. The total limit of tables in the ACTIVE state is 250.
+ * The number of concurrent table requests (cumulative number of tables in the CREATING, DELETING or UPDATING state) exceeds the maximum allowed of 10.
  */
 class LimitExceededException extends DynamoDbException {}